# src/coordination/contracts/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home23 Connected Agents API",
    "version": "1",
    "description": "Minimal contract-only draft. It defines no router or runtime implementation."
  },
...
  "servers": [
    {"url": "http://127.0.0.1:{port}", "description": "Explicit loopback listener", "variables": {"port": {"default": "5050"}}},
    {"url": "https://{vpnHost}", "description": "Authenticated Tailscale/VPN route", "variables": {"vpnHost": {"default": "home23.invalid"}}}
  ],
  "security": [{"bearerSession": []}],
  "paths": {
...
    "/api/v1/chess/games": {"get": {"operationId": "listChessGames","description": "House session authentication; game and position access require active channel membership. Chess writes require Idempotency-Key. Stream snapshots omit moves; GET returns full history.","parameters": [],"responses": {"200": {"description": "Chess response","content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/chessGameList"}}}},"default": {"$ref": "#/components/responses/Error"}}},"post": {"operationId": "createChessGame","description": "House session authentication; game and position access require active channel membership. Chess writes require Idempotency-Key. Stream snapshots omit moves; GET returns full history.","parameters": [{"name": "Idempotency-Key","in": "header","required": true,"schema": {"type": "string","minLength": 16,"maxLength": 128}}],"responses": {"201": {"description": "Chess response","content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/chessGameResponse"}}}},"default": {"$ref": "#/components/responses/Error"}},"requestBody": {"required": true,"content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/chessCreateRequest"}}}}}},
    "/api/v1/chess/games/{gameId}": {"get": {"operationId": "getChessGame","description": "House session authentication; game and position access require active channel membership. Chess writes require Idempotency-Key. Stream snapshots omit moves; GET returns full history.","parameters": [{"name": "gameId","in": "path","required": true,"schema": {"type": "string"}}],"responses": {"200": {"description": "Chess response","content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/chessGameResponse"},"x-canonical-fixture": "chess-game"}}},"default": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/chess/games/{gameId}/moves": {"post": {"operationId": "moveChessPiece","description": "House session authentication; game and position access require active channel membership. Chess writes require Idempotency-Key. Stream snapshots omit moves; GET returns full history.","parameters": [{"name": "gameId","in": "path","required": true,"schema": {"type": "string"}},{"name": "Idempotency-Key","in": "header","required": true,"schema": {"type": "string","minLength": 16,"maxLength": 128}}],"responses": {"200": {"description": "Chess response","content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/chessGameResponse"}}}},"default": {"$ref": "#/components/responses/Error"}},"requestBody": {"required": true,"content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/chessMoveRequest"}}}}}},
    "/api/v1/chess/games/{gameId}/control": {"post": {"operationId": "controlChessGame","description": "House session authentication; game and position access require active channel membership. Chess writes require Idempotency-Key. Stream snapshots omit moves; GET returns full history.","parameters": [{"name": "gameId","in": "path","required": true,"schema": {"type": "string"}},{"name": "Idempotency-Key","in": "header","required": true,"schema": {"type": "string","minLength": 16,"maxLength": 128}}],"responses": {"200": {"description": "Chess response","content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/chessGameResponse"}}}},"default": {"$ref": "#/components/responses/Error"}},"requestBody": {"required": true,"content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/chessControlRequest"}}}}}},
    "/api/v1/chess/games/{gameId}/pgn": {"get": {"operationId": "exportChessPgn","description": "House session authentication; game and position access require active channel membership. Chess writes require Idempotency-Key. Stream snapshots omit moves; GET returns full history.","parameters": [{"name": "gameId","in": "path","required": true,"schema": {"type": "string"}}],"responses": {"200": {"description": "Chess response","content": {"application/x-chess-pgn": {"schema": {"type": "string"}}}},"default": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/chess/games/{gameId}/stream": {"get": {"operationId": "streamChessGame","description": "House session authentication; game and position access require active channel membership. Chess writes require Idempotency-Key. Stream snapshots omit moves; GET returns full history.","parameters": [{"name": "gameId","in": "path","required": true,"schema": {"type": "string"}}],"responses": {"200": {"description": "Chess response","content": {"text/event-stream": {"schema": {"$ref": "./schema.json#/$defs/chessSnapshot"},"x-canonical-fixture": "chess-snapshot"}}},"default": {"$ref": "#/components/responses/Error"}}}},
    "/api/v1/chess/positions": {"get": {"operationId": "listChessPositions","description": "House session authentication; game and position access require active channel membership. Chess writes require Idempotency-Key. Stream snapshots omit moves; GET returns full history.","parameters": [],"responses": {"200": {"description": "Chess response","content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/chessPositionList"}}}},"default": {"$ref": "#/components/responses/Error"}}},"post": {"operationId": "saveChessPosition","description": "House session authentication; game and position access require active channel membership. Chess writes require Idempotency-Key. Stream snapshots omit moves; GET returns full history.","parameters": [{"name": "Idempotency-Key","in": "header","required": true,"schema": {"type": "string","minLength": 16,"maxLength": 128}}],"responses": {"201": {"description": "Chess response","content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/chessPositionResponse"}}}},"default": {"$ref": "#/components/responses/Error"}},"requestBody": {"required": true,"content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/chessPositionRequest"}}}}}},
    "/api/v1/chess/positions/{positionId}": {"get": {"operationId": "getChessPosition","description": "House session authentication; game and position access require active channel membership. Chess writes require Idempotency-Key. Stream snapshots omit moves; GET returns full history.","parameters": [{"name": "positionId","in": "path","required": true,"schema": {"type": "string"}}],"responses": {"200": {"description": "Chess response","content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/chessPositionResponse"},"x-canonical-fixture": "chess-position"}}},"default": {"$ref": "#/components/responses/Error"}}}}
  },
  "components": {
    "securitySchemes": {"bearerSession": {"type": "http", "scheme": "bearer"}},
...
    "responses": {
      "Error": {"description": "Safe stable error envelope", "headers": {"X-Request-ID": {"schema": {"type": "string"}}, "X-Correlation-ID": {"schema": {"type": "string"}}}, "content": {"application/json": {"schema": {"$ref": "./schema.json#/$defs/errorFixture"}, "x-canonical-fixture": "error"}}}
    }
  }
}